Remove dt_convergence from mismip+ group, add smoketest
The MISMIP+ test group was added in this branch and has only a
dt_convergence_test case.  Now that I have created a calving_dt_convergence
test group, there is no need to have a MISMIP+ test case for that
purpose.

This commit deletes the dt_convergence_test module and test from the
MISMIP+ test group.  It then adds a smoke test so that the MISMIP+ test
group can actually be used.  The smoke test just uses the pre-built
initial condition file and runs for 5 years with the FO solver.  It
takes about a minute and a half on Badger.  This test can be modified
or replaced with something as needs require.
