+++ compass/landice/tests/mismipplus/namelist.landice
@@ -1,18 +1,18 @@
 config_start_time = '0000-01-01_00:00:00'
-config_run_duration = '0010-00-00_00:00:00'
+config_run_duration = '0005-00-00_00:00:00'
 config_adaptive_timestep = .true.
 config_adaptive_timestep_CFL_fraction = 0.9
-config_adaptive_timestep_include_calving = .true.
+config_adaptive_timestep_include_calving = .false.
 config_adaptive_timestep_calvingCFL_fraction = 1.0
 config_block_decomp_file_prefix = 'graph.info.part.'
-config_velocity_solver = 'none'
+config_velocity_solver = 'FO'
 config_thickness_advection = 'fo'
 config_tracer_advection = 'none'
 config_thermal_solver = 'none'
 config_temperature_init = 'file'
 config_dynamic_thickness = 0.001
 
-config_calving = 'von_Mises_stress'
+config_calving = 'none'
 config_floating_von_Mises_threshold_stress_source = 'scalar'
 config_floating_von_Mises_threshold_stress = 20000.0
 config_calving_specified_source = 'const'
