+++ compass/landice/tests/mismipplus/run_model.py
@@ -16,7 +16,7 @@ class RunModel(Step):
         twice, the second time with ``namelist.landice.rst`` and
         ``streams.landice.rst``
     """
-    def __init__(self, test_case, calv_dt_frac=None, name='run_model', subdir=None,
+    def __init__(self, test_case, name='run_model', subdir=None,
                  cores=1, min_cores=None, threads=1, suffixes=None):
         """
         Create a new test case
@@ -25,9 +25,6 @@ class RunModel(Step):
         ----------
         test_case : compass.TestCase
             The test case this step belongs to
-
-        calv_dt_frac : float, optional
-            the value to use for calving dt fraction
 
         name : str, optional
             the name of the test case
@@ -77,12 +74,6 @@ class RunModel(Step):
             self.add_streams_file(
                 'compass.landice.tests.mismipplus', 'streams.landice',
                 out_name='streams.{}'.format(suffix))
-
-            if calv_dt_frac is not None:
-                options = {'config_adaptive_timestep_calvingCFL_fraction':
-                               f"{calv_dt_frac}"}
-                self.add_namelist_options(options=options,
-                                          out_name='namelist.{}'.format(suffix))
 
         self.add_input_file(filename='albany_input.yaml',
                             package='compass.landice.tests.mismipplus',
